Create Legal Compliance
I propose that we create a Folder called Legal Compliance that will eventually include sections that explain if Grin complies with the laws of the various countries. I believe that there will be more legal contributors for country specific rules which is needed for wide-scale adoption. I am going to start the docs/`United States Legal Compliance`. The United States docs will generally deal with federal law. I do not expect that we will focus on the state laws because they will be beyond the scope of our abilities. However, if congress passes a rule making cryptocurrency regulation a state by state issue, which I doubt, then it may become necessary.