Add instructions for creating a wallet service
Add a brand new wiki page in `extra-documents/` to guide the user
through creating a new Grin wallet service. This guide also covers
handling wallet password authentication as well as verifying that the
wallet is reachable.

Also a few links to this new guide has been added to the `receive.md`
quickstart page.